+++ alex_nb_web/src/main/java/nadrabank/dao/AssetHistoryDaoImpl.java
@@ -7,7 +7,6 @@ import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -47,22 +46,22 @@ public class AssetHistoryDaoImpl implements AssetHistoryDao {
         return true;
     }
     @Override
-    public BigDecimal getFirstAccPrice(Long assId) {
+    public Integer getFirstAccPrice(Long assId) {
         Query query = factory.getCurrentSession().createQuery("select ah.acceptPrice from AssetHistory ah where ah.id=:assId and ah.acceptPrice is not null ORDER BY ah.changeDate ASC ");
         query.setParameter("assId", assId);
         try {
-            return (BigDecimal) query.list().get(0);
+            return (int) query.list().get(0);
         }
         catch(IndexOutOfBoundsException e){
             return null;
         }
     }
     @Override
-    public BigDecimal getLastAccPrice(Long assId) {
+    public Integer getLastAccPrice(Long assId) {
         Query query = factory.getCurrentSession().createQuery("select ah.acceptPrice from AssetHistory ah where ah.id=:assId and ah.acceptPrice is not null ORDER BY ah.changeDate DESC ");
         query.setParameter("assId", assId);
         try {
-            return (BigDecimal) query.list().get(0);
+            return (int) query.list().get(0);
         }
         catch(IndexOutOfBoundsException e){
             return null;
@@ -81,12 +80,12 @@ public class AssetHistoryDaoImpl implements AssetHistoryDao {
     }
 
     @Override
-    public BigDecimal getAccPriceByLotIdHistory(Long assetId, Long lotId){
+    public Integer getAccPriceByLotIdHistory(Long assetId, Long lotId){
         Query query = factory.getCurrentSession().createQuery("select ah.acceptPrice from AssetHistory ah where ah.idKey in (SELECT max(a.idKey) from AssetHistory a where a.id=:assetId and a.lotId=:lotId)");
         query.setParameter("assetId", assetId);
         query.setParameter("lotId", lotId);
         try {
-            return (BigDecimal) query.list().get(0);
+            return (int) query.list().get(0);
         }
         catch(IndexOutOfBoundsException e){
             return null;
